use custom hook to manage modal state
